requirements.txt
fastmcp
requests
httpx
typing